Load Favorite weather details

+++ app/src/main/java/com/omarinc/weather/currentHomeWeather/view/MyHourAdapter.kt
@@ -23,7 +23,7 @@ class MyHourAdapter : ListAdapter<TodayForecast, MyHourAdapter.ViewHolder>(
 
         with(holder.binding) {
             tvTimeHours.text = currentForecast.time
-            tvDegreeHours.text = String.format("%s°C", currentForecast.temp.toInt().toString())
+            tvDegreeHours.text = String.format("%s°C", currentForecast.temp.toString())
             setIcon(currentForecast.icon, ivStatusIconHours)
         }
     }
